phonegap support; add hesperian_mobile_init.js as a place for jquery mobile initialization.
phongap - we'll create the html in the phonegap/www directory expected by the phonegap xcode project. It's the same as the regular html exept that it omits the manifest file (since it's all local on the phone), and includes phonegap.js. Note that this is an xcode 3.2.6. xcode 4 is out, but still has some stability issues (and it's a bit of a pain to maintain both versions of xcode on the same machine - I've not yet moved to 4).

As to the hesperian_mobile_init.js, we'll include that before including jquery mobile. Some of the initialization of JQM needs to be before JQM loads. Anything that needs to be after, we can put in hesperian_mobile.js