Phonegap build improvements: phonegap plugin to disable bounce scrolling, stub art
We'll address https://github.com/saqimtiaz/hesperian-mhealth-pregnancy-warning-signs/issues/25 "Disable elastic scrolling".
To do this, we need to dig down to the UIScrollView owned by the UIWebView displaying our html, and set it to bounces = NO.
Phone gap provides a plugin architecture whereby we can add additional functionality, callable from javascript. We'll create a HesperianMobile.init function which we call in our js, which calles out to ObjectiveC in the phonegap iOS project.

In the ObjectiveC code, we recurse down to find the UIScrollView, and disable bounces. Note that this assumes that the UIWebView scrolls via a UIScrollview, which seems to be true as of iOS 4.0. Earlier versions do things differently - (tested iOS 3.2.1), so bounce will still be present there.